Add command-line -prefix option.
This allows a filename prefix for the generated Parser/Scanner files.
The text token "$PREFIX$" is replaced during the copy frames process.

Example,
    Coco -prefix Foo -namespace Project::Bar myGrammar.atg
generates
    FooParser.h, FooParser.cpp, FooScanner.h, FooScanner.cpp
    in the namespace Project { namespace Bar { ...

The name of the include 'protection' also gets adjusted accordingly.
From

    #ifndef COCO_$PREFIX$SCANNER_H__
    #define COCO_$PREFIX$SCANNER_H__

to

    #ifndef COCO_FooSCANNER_H__
    #define COCO_FooSCANNER_H__

The primary motivation for all of this is when combining different
scanners/parsers in a library. Even with different namespaces, we have
to ensure that the filenames don't collide.

Note that the generated classes themselves retain the names 'Parser'
and 'Scanner' within their namespaces ... at least some things remain
constant.