Add ability to hook in user error handling into the Parser.frame
The Parser now has an optional construct parameter:

    Parser(Scanner* scan, Errors* err = 0);

If a pointer to an 'Errors' class (or derived class) is provided, it
will be used. This pointer will NOT be deallocated when the Parser is
destroyed.

If a null pointer, or pointer is provided, the standard 'Errors' class
will be allocated and used ... and deleted when the Parser is
destroyed.