Add method to Buffer classes to distinguish UTF8 vs. non-UTF8
On windows systems we can (probably) rely on the existence of a
byte-order-mark to signal UTF8 encoding vs. unencoded files. If this
fails, we have the same situation as on non-windows systems.

If a UTF8 encoded file is read directly with Buffer class (instead of
with the UTF8Buffer class), the wide-character representation may make
things look like other unicode characters if we aren't careful.

For example, the Euro symbol (€) U+20AC appears as the following UTF8
byte sequence: (â²¬) 0xE2 0x82 0xAC. If each character is treated as
unicode and outputted as UTF8, the final output will get out of
control.