Avoid spurious leading whitespace and single trailing blank on copied text
The leading space can arise when surrounding tokens are used to define
the content. This is what is done when copying the [code] .. [/code]
block to preserve initial comments in the chunk.

The single trailing blank often arises in the semantic action. For
readability it may be preferable to avoid (.code.) in favour of using
the (. code .) form instead. This however adds a spurious trailing
blank that not only looks slightly messy, but can also cause other
annoyances with git etc.