Add initial version of a preprocessor for the frame files.
This should allow leaner or specialized final code that can be
selected when the grammar is compiled.

The first use would be to explicity specify if UTF8 encoding should be
used (on non-windows systems). A later use might be to selective
disable chunks of code - eg, if only the STL streams are to be used in
an application, the stdio bits could be disabled.