ENH: find frames in a share directory relative to the executable
If *--prefix* (not *--exec-prefix*) is used when configuring and
installing coco-cpp (on Unix-style systems), the files will be
installed, for example, in these locations

    /usr/bin/coco-cpp
    /usr/share/coco-cpp/Parser.frame
    /usr/share/coco-cpp/Scanner.frame

The frame files can then be found in a share directory relative to the
executable, without needing to specify *-frames* separately.

ENH: Report which frame files were found and used for the grammar.